test: Updated unittests
