+++ test/common/test_utils_fonts.py
@@ -1,4 +1,4 @@
-from PIL.ImageFont import ImageFont
+from PIL.ImageFont import ImageFont, FreeTypeFont
 
 from doctr.utils.fonts import get_font
 
@@ -8,4 +8,4 @@ def test_get_font():
     # Attempts to load recommended OS font
     font = get_font()
 
-    assert isinstance(font, ImageFont)
+    assert isinstance(font, (ImageFont, FreeTypeFont))
