+++ test/tensorflow/test_models_tf.py
@@ -2,6 +2,7 @@ import pytest
 import numpy as np
 
 from doctr import models
+from doctr.models.predictor import OCRPredictor
 from doctr.io import Document, DocumentFile
 from test_models_detection_tf import test_detectionpredictor, test_rotated_detectionpredictor
 from test_models_recognition_tf import test_recognitionpredictor
@@ -11,12 +12,12 @@ def test_ocrpredictor(
     mock_pdf, test_detectionpredictor, test_recognitionpredictor, test_rotated_detectionpredictor  # noqa: F811
 ):
 
-    predictor = models.OCRPredictor(
+    predictor = OCRPredictor(
         test_detectionpredictor,
         test_recognitionpredictor
     )
 
-    r_predictor = models.OCRPredictor(
+    r_predictor = OCRPredictor(
         test_rotated_detectionpredictor,
         test_recognitionpredictor,
         rotated_bbox=True
@@ -49,4 +50,4 @@ def test_zoo_models(det_arch, reco_arch):
     # Model
     predictor = models.ocr_predictor(det_arch, reco_arch, pretrained=True)
     # Output checks
-    assert isinstance(predictor, models.OCRPredictor)
+    assert isinstance(predictor, OCRPredictor)
